Delete inaccurate 'change password' display manager needle
This needle is just the same as standard display manager but not checking on
the username. So it will always match in all cases and is at least confusing
why the 'change password' needle is matching in any default scenario, e.g. see
https://openqa.suse.de/tests/515515#step/first_boot/1 . If it is necessary to
have it, it should be recreated more specifically.