Delete invalid bug detection needle for bsc#993249
bsc#993249 is marked as fixed and was previously verified. The older
partitioning_lvm-partitioning-encrypt-broke-existing-20160823 needle was used
to detect a partition being formatted with "unknown" which would yield an
unusable system. The needle
partitioning_lvm-partitioning-encrypt-broke-existing-20160907 is erroneously
matching and causing a soft failure to the already fixed bug to be recorded
although the bug is VERIFIED FIXED *and* the test actually continues just fine
to boot into a usable system, e.g. see
https://openqa.suse.de/tests/603779#step/first_boot/4