Verify installation overview is loaded
During execution on slow system loading of overview may take time,
whereas we start pressing tab key until get to the software section. It
is in the top, and if we press tab couple of times before, we won't be
able to reach software section in 20 presses for some installations.
For that we need one new needle and adding tags to needles that can be
reused.

See poo#12322.