Grub boot menu with pxe_boot
With pxe_boot menu we have different appearance of grub, so wait_boot
method doesn't work. Here we skip PXE boot menu, but it will boot to
hard disk by default, hence is acceptable as we need this for
softfailure in one of the tests only. Additional checks may be added
later on.