Add relevant tags to new grub needles
In MR#407 we created new needles for grub, after discussion we decided
to rename the files again and add couple of relevant tags to unregister
irrelevant needles for specific tests. There will be another PR for
main.pm to unregister these new needles in case PXEBOOT variable is not
set.