Needles for changes in yast2_ui tests
We introduce changes to yast2_ui tests for openSUSE and once in a while
add new needles. As test suite is relatively stable on SLE, no changes
are required for it, but as we introduce additional steps, we need to
introduce new needles here as well.

See poo#19922 and blocking issues.