Fix Release Notes test issues with SLE-HA - Again!
The last commit was ok on my lab but not on o.s.d.

This is due to the fact that we need to add SLES product
with SLEHA, and so we need a needles with the tag
'release-notes-sles' and not 'release-notes-sle'.

This commit fix this by adding the correct tag and renaming
the needle.

Failed run: https://openqa.suse.de/tests/1424672#step/releasenotes/9

Note: as my verification run was only ok on my lab, I would
run the verification directly on o.s.d. Only SLE-HA seems to be
impacted by this from now.