Prevent erroneous match on text-login prompt by using only tty-specific needles
There should be a tty-specific needle in all cases for "select_console" calls
so we should not need the generic "text-login" needles which can only confuse
tests as a text login prompt might show up for a wrong tty, the tty is changed
while the username is typed and when the final tty shows up the username was
never typed. This commit deletes the ambiguous needles assuming that there are
tty-specific needles as alternatives, e.g. "tty2-selected" or "tty6-selected".

Related progress issue: https://progress.opensuse.org/issues/33097