Remove too generic text-login needle
This is short-term solution to the issue. We should assert that we are
on the right tty and not on any. This is risky change though, so cannot
be easily introduced. As of now we can create correct needles. I will
propose PR with changes to os-autoinst, so we could force tty selection
assertion.

See [poo#32926](https://progress.opensuse.org/issues/32926).