Fix installation on zkvm due to wrong needle match
The needles were added as an attempt to fix 'No login prompt issue, but
after investigation it was found out, that the issue is in the product
itself - bsc#1111524. Also, the needle matched wrong step and caused fail
on the start of installation, as it matched any black screen.

The commit removes the needles as invalid.

[poo#41861](https://progress.opensuse.org/issues/41861)