Handle sles4sap selection in textmode
Starting from Build 70.3, an additional product (RT) is shown on Product
Selection screen. The key combination for the sles4sap is changed from
'alt+i' to 'alt+x' in textmode that is also changed the highlighted
character for the sles4sap.

The commit adds the needles for the actual UI.