Correct partitioning_firstdisk-* matching areas
partitioning_firstdisk-select-hard-disks-one-selected-20180601:
Make matching area taller so as to include two checkboxes rather than
one, lower one unselected.
Purpose: forestall possible bogus assert_screen matches for this tag in

    lib/partition_setup.pm, sub select_first_hard_disk()

which would erroneously proceed to skip the subsequent assert_and_click
even if a second disk in need of unselection is present.

partitioning_firstdisk-select-hard-disks-sdb-20180420:
Obviously it is the lower rather than the upper matching area which
is intended for

    lib/partition_setup.pm, sub select_first_hard_disk()
    assert_and_click 'hard-disk-dev-sdb-selected' if ...

therefore it is this matching area which has to come last in the .json.